+++ src/assets/index.ts
@@ -1,23 +1,15 @@
 import Loader from './blobloader.gif';
 import BlobDivider from './blackblob1.png';
-import Slots1 from './content/slots-1.png';
-import Slots2 from './content/slots-2.png';
-import Slots3 from './content/slots-3.png';
-import Zebra1 from './content/zebra-1.png';
-import Zebra2 from './content/zebra-2.png';
-import Zebra3 from './content/zebra-3.png';
-import Noise1 from './noise-01.png';
+import noise_01 from './noise-01.png';
 
 const AssetMap = {
   BlobDivider,
   Loader,
-  Slots1,
-  Slots2,
-  Slots3,
-  Zebra1,
-  Zebra2,
-  Zebra3,
-  Noise1
+  noise_01
+}
+
+export function getUrl(name: string) {
+  return new URL(`./content/${name}`, import.meta.url).href;
 }
 
 export default AssetMap;
