+++ src/components/content/content_home.tsx
