+++ vite.config.ts
@@ -4,5 +4,5 @@ import react from '@vitejs/plugin-react'
 // https://vitejs.dev/config/
 export default defineConfig({
   base: '/portfolio/',
-  plugins: [react()],
+  plugins: [react()]
 })
